controllers\opml: split into separate controllers
This will allow us to only inject the necessary classes.

No other changes were made.
